Create DISCLAIMER.md
GS1®, under its IP Policy (https://www.gs1.org/standards/ip), seeks to avoid uncertainty regarding intellectual property claims by requiring the participants in the Work Group that developed the GS1 Standards to agree to grant to GS1 members a royalty-free licence or a RAND licence to Necessary Claims, as that term is defined in the GS1 IP Policy. Furthermore, attention is drawn to the possibility that an implementation of one or more features of the GS1 Standards may be the subject of a patent or other intellectual property right that does not involve a Necessary Claim. Any such patent or other intellectual property right is not subject to the licencing obligations of GS1. Moreover, the agreement to grant licences provided under the GS1 IP Policy does not include IP rights and any claims of third parties who were not participants in the Work Group. ​

Accordingly, GS1 recommends that any organization developing an implementation designed to be in conformance with GS1 Standards should determine whether there are any patents that may encompass a specific implementation that the organisation is developing in compliance with the relevant GS1 Standards and whether a licence under a patent or other intellectual property right is needed. Such a determination of a need for licencing should be made in view of the details of the specific system designed by the organisation in consultation with their own patent counsel. ​

THE GS1 STANDARDS ARE PROVIDED “AS IS” WITH NO WARRANTIES WHATSOEVER, INCLUDING ANY WARRANTY OF MERCHANTABILITY, NONINFRINGMENT, FITNESS FOR PARTICULAR PURPOSE, OR ANY WARRANTY OTHER WISE ARISING OUT OF THE GS1 STANDARDS. GS1 disclaims all liability for any damages arising from use or misuse of the GS1 Standards, whether special, indirect, consequential, or compensatory damages, and including liability for infringement of any intellectual property rights, relating to use of information in or reliance upon the GS1 Standards. ​

GS1 retains the right to make changes to the GS1 Standards at any time, without notice. GS1 makes no warranty for the use of the GS1 Standards and assumes no responsibility for any errors which may appear in the GS1 Standards, nor does it make a commitment to update the information contained herein. ​

GS1 and the GS1 logo are registered trademarks of GS1 AISBL.